% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_NNT.R
\name{compute_NNT}
\alias{compute_NNT}
\title{Compute Number Needed to treat}
\usage{
compute_NNT(e, n, alpha)
}
\arguments{
\item{e}{a bivariate numeric vector of counts of events in the two treatment groups}

\item{n}{a bivariate numeric vector of size of the two treatment groups}

\item{alpha}{a numeric valie between 0 and 1 so as to obtain 1-alpha confidence interval}
}
\value{
estimate of absolute risk reduction ARR, number needed to treat NNT and 1-alpha confidence intervals
}
\description{
Compute Number Needed to treat
}
\details{
confidence
}
